Rename widget to object
