--- src/analysis/widget.rs
@@ -1,9 +1,0 @@
-use env::Env;
-use gobjects::GObject;
-use super::object;
-
-pub fn new(env: &Env, obj: &GObject) -> Option<object::Info> {
-    let info = object::new(env, obj);
-
-    info
-}
