+++ src/codegen/objects.rs
@@ -13,7 +13,7 @@ pub fn generate(env: &Env, root_path: &Path, mod_rs: &mut Vec<String>, traits: &
         }
 
         info!("Analyzing {:?}", obj.name);
-        let class_analysis = match analysis::widget::new(env, obj) {
+        let class_analysis = match analysis::object::new(env, obj) {
             Some(info) => info,
             None => {
                 warn!("Class {} not found.", obj.name);
@@ -25,9 +25,9 @@ pub fn generate(env: &Env, root_path: &Path, mod_rs: &mut Vec<String>, traits: &
         info!("Generating file {:?}", path);
 
         save_to_file(path, env.config.make_backup,
-            |w| super::widget::generate(w, env, &class_analysis));
+            |w| super::object::generate(w, env, &class_analysis));
 
         let mod_name = module_name(split_namespace_name(&class_analysis.full_name).1);
-        super::widget::generate_reexports(env, &class_analysis, &mod_name, mod_rs, traits);
+        super::object::generate_reexports(env, &class_analysis, &mod_name, mod_rs, traits);
     }
 }
